+++ log/driver.go
@@ -1,7 +1,6 @@
 package log
 
 import (
-	"fmt"
 	"time"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
@@ -9,18 +8,11 @@ import (
 
 // Driver makes trace.Driver with internal logging
 func Driver(log Logger, details trace.Details) trace.Driver {
-	log, ok := log.WithName(`driver`).(Logger)
-	if !ok {
-		panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-	}
+	log = log.WithName(`driver`)
 	t := trace.Driver{}
 	// nolint: nestif
 	if details&trace.DriverNetEvents != 0 {
-		// nolint: govet
-		log, ok := log.WithName(`net`).(Logger)
-		if !ok {
-			panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-		}
+		log = log.WithName(`net`)
 		t.OnNetRead = func(info trace.NetReadStartInfo) func(trace.NetReadDoneInfo) {
 			address := info.Address
 			log.Tracef(`read start {address:"%s"}`,
@@ -110,11 +102,7 @@ func Driver(log Logger, details trace.Details) trace.Driver {
 	}
 	// nolint: nestif
 	if details&trace.DriverCoreEvents != 0 {
-		// nolint: govet
-		log, ok := log.WithName(`core`).(Logger)
-		if !ok {
-			panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-		}
+		log = log.WithName(`core`)
 		t.OnInit = func(info trace.InitStartInfo) func(trace.InitDoneInfo) {
 			endpoint := info.Endpoint
 			database := info.Database
@@ -304,11 +292,7 @@ func Driver(log Logger, details trace.Details) trace.Driver {
 		}
 	}
 	if details&trace.DriverDiscoveryEvents != 0 {
-		// nolint: govet
-		log, ok := log.WithName(`discovery`).(Logger)
-		if !ok {
-			panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-		}
+		log = log.WithName(`discovery`)
 		t.OnDiscovery = func(info trace.DiscoveryStartInfo) func(trace.DiscoveryDoneInfo) {
 			log.Debugf(`discover start`)
 			start := time.Now()
@@ -328,11 +312,7 @@ func Driver(log Logger, details trace.Details) trace.Driver {
 		}
 	}
 	if details&trace.DriverClusterEvents != 0 {
-		// nolint: govet
-		log, ok := log.WithName(`cluster`).(Logger)
-		if !ok {
-			panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-		}
+		log = log.WithName(`cluster`)
 		t.OnClusterGet = func(info trace.ClusterGetStartInfo) func(trace.ClusterGetDoneInfo) {
 			log.Tracef(`get start`)
 			start := time.Now()
@@ -422,11 +402,7 @@ func Driver(log Logger, details trace.Details) trace.Driver {
 		}
 	}
 	if details&trace.DriverCredentialsEvents != 0 {
-		// nolint: govet
-		log, ok := log.WithName(`credentials`).(Logger)
-		if !ok {
-			panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-		}
+		log = log.WithName(`credentials`)
 		t.OnGetCredentials = func(info trace.GetCredentialsStartInfo) func(trace.GetCredentialsDoneInfo) {
 			log.Tracef(`get start`)
 			start := time.Now()