+++ log/table.go
@@ -1,7 +1,6 @@
 package log
 
 import (
-	"fmt"
 	"time"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
@@ -10,18 +9,10 @@ import (
 // Table makes trace.Table with internal logging
 // nolint: gocyclo
 func Table(log Logger, details trace.Details) trace.Table {
-	log, ok := log.WithName(`table`).(Logger)
-	if !ok {
-		panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-	}
+	log = log.WithName(`table`)
 	t := trace.Table{}
-	// nolint: nestif
 	if details&trace.TablePoolRetryEvents != 0 {
-		// nolint: govet
-		log, ok := log.WithName(`retry`).(Logger)
-		if !ok {
-			panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-		}
+		log = log.WithName(`retry`)
 		// nolint: lll
 		t.OnPoolRetry = func(info trace.PoolRetryStartInfo) func(info trace.PoolRetryInternalInfo) func(trace.PoolRetryDoneInfo) {
 			idempotent := info.Idempotent
@@ -63,11 +54,7 @@ func Table(log Logger, details trace.Details) trace.Table {
 	}
 	// nolint: nestif
 	if details&trace.TableSessionEvents != 0 {
-		// nolint: govet
-		log, ok := log.WithName(`session`).(Logger)
-		if !ok {
-			panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-		}
+		log = log.WithName(`session`)
 		if details&trace.TableSessionLifeCycleEvents != 0 {
 			t.OnSessionNew = func(info trace.SessionNewStartInfo) func(trace.SessionNewDoneInfo) {
 				log.Tracef(`create start`)
@@ -136,17 +123,9 @@ func Table(log Logger, details trace.Details) trace.Table {
 			}
 		}
 		if details&trace.TableSessionQueryEvents != 0 {
-			// nolint: govet
-			log, ok := log.WithName(`query`).(Logger)
-			if !ok {
-				panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-			}
+			log = log.WithName(`query`)
 			if details&trace.TableSessionQueryInvokeEvents != 0 {
-				// nolint: govet
-				log, ok := log.WithName(`invoke`).(Logger)
-				if !ok {
-					panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-				}
+				log = log.WithName(`invoke`)
 				t.OnSessionQueryPrepare = func(info trace.SessionQueryPrepareStartInfo) func(trace.PrepareDataQueryDoneInfo) {
 					session := info.Session
 					query := info.Query
@@ -216,11 +195,7 @@ func Table(log Logger, details trace.Details) trace.Table {
 				}
 			}
 			if details&trace.TableSessionQueryStreamEvents != 0 {
-				// nolint: govet
-				log, ok := log.WithName(`stream`).(Logger)
-				if !ok {
-					panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-				}
+				log = log.WithName(`stream`)
 				// nolint: lll
 				t.OnSessionQueryStreamExecute = func(info trace.SessionQueryStreamExecuteStartInfo) func(trace.SessionQueryStreamExecuteDoneInfo) {
 					session := info.Session
@@ -284,11 +259,7 @@ func Table(log Logger, details trace.Details) trace.Table {
 			}
 		}
 		if details&trace.TableSessionTransactionEvents != 0 {
-			// nolint: govet
-			log, ok := log.WithName(`transaction`).(Logger)
-			if !ok {
-				panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-			}
+			log = log.WithName(`transaction`)
 			// nolint: lll
 			t.OnSessionTransactionBegin = func(info trace.SessionTransactionBeginStartInfo) func(trace.SessionTransactionBeginDoneInfo) {
 				session := info.Session
@@ -377,11 +348,7 @@ func Table(log Logger, details trace.Details) trace.Table {
 	}
 	// nolint: nestif
 	if details&trace.TablePoolEvents != 0 {
-		// nolint: govet
-		log, ok := log.WithName(`pool`).(Logger)
-		if !ok {
-			panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-		}
+		log = log.WithName(`pool`)
 		if details&trace.TablePoolLifeCycleEvents != 0 {
 			t.OnPoolInit = func(info trace.PoolInitStartInfo) func(trace.PoolInitDoneInfo) {
 				log.Infof(`initialize start`)
@@ -412,11 +379,7 @@ func Table(log Logger, details trace.Details) trace.Table {
 			}
 		}
 		if details&trace.TablePoolSessionLifeCycleEvents != 0 {
-			// nolint: govet
-			log, ok := log.WithName(`session`).(Logger)
-			if !ok {
-				panic(fmt.Sprintf("%T.WithName() returns interface which not casted to Logger interface", log))
-			}
+			log = log.WithName(`session`)
 			t.OnPoolSessionNew = func(info trace.PoolSessionNewStartInfo) func(trace.PoolSessionNewDoneInfo) {
 				log.Tracef(`create start`)
 				start := time.Now()